requirements.txt
streamlit>=1.31.0
google-genai>=0.3.0